+++ src/index.js
@@ -1,4 +1,4 @@
 export { wrapStream }                       from './stream.js'
 export { wrapWebSocket, PqTlsWebSocket }   from './websocket.js'
 export { initiatorHandshake, responderHandshake } from './handshake.js'
-export { KxcoPqTlsError }                  from './errors.js'
+export { KxcoPqTlsError, ERR_HANDSHAKE_TIMEOUT } from './errors.js'